requirements.txt
asgiref==3.2.4
Django==3.0.4
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.1
psycopg2==2.8.4
gunicorn==20.0.4
django-storages==1.9.1
google-cloud-storage==1.27.0
